Add class for adding new books
