+++ main.js
@@ -1,21 +1,17 @@
-import Book from './Book.js';
-import Remove from './remove.js';
+import Remove from './module/remove.js';
+import Add from './module/add.js';
 
+const add = new Add();
 const remove = new Remove();
-const bookTitle = document.querySelector('#bookname');
-const bookAuthor = document.querySelector('#authorname');
 const createNew = document.querySelector('#createNewButton');
 const bookContainer = document.querySelector('#book-container');
 let allBooks = [];
-let id;
 if (localStorage.getItem('bookList') === null) {
   localStorage.setItem('bookList', []);
 }
 if (localStorage.getItem('id') === null) {
   localStorage.setItem('id', JSON.stringify(0));
-  id = JSON.parse(localStorage.getItem('id'));
 }
-
 const refreshDOM = () => {
   allBooks = JSON.parse(localStorage.getItem('bookList'));
   allBooks.forEach((book) => {
@@ -41,27 +37,9 @@ const refreshDOM = () => {
   });
 };
 window.onload = refreshDOM;
-
 createNew.addEventListener('click', (e) => {
   e.preventDefault();
-  const name = bookTitle.value;
-  bookTitle.value = '';
-  const author = bookAuthor.value;
-  bookAuthor.value = '';
-  if (!(name.length < 3 || author.length < 3)) {
-    id = JSON.parse(localStorage.getItem('id'));
-    id += 1;
-    localStorage.setItem('id', JSON.stringify(id));
-    const newBook = new Book(id, name, author);
-    if (localStorage.getItem('bookList').length !== 0) {
-      allBooks = JSON.parse(localStorage.getItem('bookList'));
-    } else {
-      allBooks = [];
-    }
-
-    allBooks.unshift(newBook);
-    localStorage.setItem('bookList', JSON.stringify(allBooks));
-    bookContainer.innerHTML = '';
-    refreshDOM();
-  }
+  add.add();
+  bookContainer.innerHTML = '';
+  refreshDOM();
 });
