+++ module/Book.js
@@ -1,9 +1,7 @@
 export default class Book {
   constructor(id, name, author) {
     this.id = id;
-
     this.name = name;
-
     this.author = author;
   }
 }
