+++ module/remove.js
